requirements_bots.txt
fastapi>=0.110.0
uvicorn[standard]>=0.27.0
httpx>=0.27.0
discord.py>=2.3.2
python-telegram-bot>=21.0
python-dotenv>=1.0.0
